+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * added `silent` param to `clear` method, [#477](https://github.com/t1m0n/air-datepicker/issues/477) 
 * fixed error when trying to call `update` method with `view` parameter on hidden calendar, [#476](https://github.com/t1m0n/air-datepicker/issues/476) 
 * fixed styles issue with time sliders, [#489](https://github.com/t1m0n/air-datepicker/issues/489) 
+* fixed English docs description, [#490](https://github.com/t1m0n/air-datepicker/issues/490) 
+* fixed typings for `formatDate` method, [#491](https://github.com/t1m0n/air-datepicker/issues/491) 
 
 ### v3.1.1
 * fixed css compilation, [#461](https://github.com/t1m0n/air-datepicker/issues/461)